correzioni

+++ Milestone 2/pipex/get_next_line/get_next_line.h
@@ -30,6 +30,6 @@ char	*ft_strchr(char *s, int c);
 char	*ft_strjoin(char *s1, char *s2);
 char	*ft_extract_line(char *str);
 char	*ft_new_str(char *str);
-size_t	ft_strlen(const char *s);
+int		ft_strlen(const char *s);
 
 #endif
